Owner displayed at top, no duplicate owner fields displayed
Owner accordion is at the top of the page and shows all available fields
except primaryID (arbitrary # not used for filtering data). Any owner
fields that are present in another file are not displayed within that
accordion - meaning if Assessment has a muni code category, the
Assessment accordion will not also contain a muni code multi-select
since the owner accordion will have that already.